refactor: remove train method (#3)
* refactor: remove train method

There is no training phase in KNN. The train() method just prepares the
data structure so that the algorithm can be applied. Do that in the
constructor instead.

BREAKING CHANGES:
The train method has been removed. Instead the arguments must be passed
to the KNN constructor.

* feat: throw if the custom distance function is not passed to load

* remove toJSON and load

They do not make sense. The internal data structure is much larger than the initial data

* eslint-fix

* rename variables

* remove unused variable